Pad rpath in Linux host binaries to make space for patching
To be able to run Linux host binaries on certain systems we patch the
rpath entry. The tool that does this (patchelf) corrupts the binary if
it needs to resize the rpath entry. This puts a bunch of no-op padding
in the rpath entry at link time so it can be patched without resizing.