Be explicit about the iOS SDK version and deployment targets.
Our minimum supported iOS version was 8.0. But the SDK we used was the first one returned by “xcodebuild —version —sdk”. Now we are explicit about that version too. This way, we can use newer features guarded by the appropriate version checks. If the Xcode installed on the host is too old or does not include the appropriate SDK, the GN step itself will terminate with a error.