Cleanup third party dependency .gitignores.
Adding new dependencies in the engine necessitates a corresponding addition of an entry to the buildroot .gitignore. This new scheme ignores everything except managed dependencies already present in the third_party directory. New managed dependencies now live in //flutter/third_party anyway so updates to this directory should be minimal (by not ignored).