Enable -Wnewline-eof
Note that as of C++11, the standard specifies that such files shall be
treated as though they include a newline at end-of-file. We add this
warning solely to match Fuchsia's clang warnings and ensure we don't
commit files without newline at EOF that will later break when rolled to
Fuchsia.