Buildroot update for Fuchsia toolchain update.
This paves the way for updating the Fuchsia toolchain to git_revision:3520297039d69381cb5776235209fc08863b9095.

The build breaks are from the introduction of new warnings that trip on translation units in our dependencies. These dependencies cannot be directly patched because they also build targets for Fuchsia but with their own toolchain versions specified in their respective buildroots. This unblocks Flutter.

There seems to have been a GN update since the last time the files I touched were formatted. So this PR contains updates to the formats of those files.