Fix OCMock static_library target
In flutter/buildroot#399, the static_library target was split into a
source_set and static_library target, but this resulted in libocmock
being built as an empty 88-byte library. We never caught the issue since
this buildroot change was never rolled to the engine.

This reverts things back to match how we build other static libraries
such as libzip and libxml.

This reverts 940dcf6e85301c049576da58b193cf2d903539a2.