Use CIPD Windows SDK to generate WinRT headers
Rather than relying on the Windows SDK that happens to be on the host
machine, we now pull down the Windows SDK from CIPD into
third_party/windows_sdk and use it directly.

cppwinrt relies on a registry entry to determine the SDK location and
doesn't currently support a mechanism for specifying this location
explicitly via an argument. Instead, we do what cppwinrt itself does and
parse out the Platform.XML file to determine the set of .winmd files to
take as input, then pass them to the tool directly.