Add build rules for family of dependencies that work with SPIRV.
This patch adds build rules for the following dependencies. These dependencies
will be added to DEPS in the engine when as this patch is rolled in:

* `glslang`: 9431c53c84c14fa9e9cd37678262ebba55c62c87
* `shaderc`: 948660cccfbbc303d2590c7f44a4cee40b66fdd6
* `spirv_cross`: 418542eaefdb609f548d25a1e3962fb69d80da63
* `spirv_tools`: 1020e394cb1267332d58497150d2b024371a8e41

The primary use case for these dependencies is in the Impeller offline pipeline
processing workflow. However, these are now necessary also to add unit-tests for
`//flutter/lib/spirv`.

Of particular note is the explicit rule for `spirv_tools`. There is already a
version of `spriv_tools` in the engine via its `SwiftShader` dependency.
However, `SwiftShader` adds this dependency directly to its sources and this
version is fairly old. That version is also incompatible with the requirements
of the other newly added dependencies. To let `SwiftShader` pick the dependency
it wants and to use the latest versions everywhere else, an explicit
`spirv_tools` dependency will be added. GN target names do not collide and
`SwiftShader` targets are used via a dynamic library.

Also, while there are no build rules for the SPIRV headers, they are necessary
and will be added to the engine DEPS.