Eliminate landmines scripts
build/landmines.py writes out a .landmines file at the root of the
buildroot. The first step of a `gclient sync` is to check for diffs
between that file and the new one, and if found, a clobber is performed.

We never actually use this functionality, and the code calls rmtree()
which (at least on macos) appears to choke on symlinks.

This deletes these scripts, and will be paired with the deletion of the
first gclient hook in the DEPS file in flutter/engine during the roll.

If we ever need this functionality, it's relatively straightforward to
add it back.

Stack trace reported on macOS 11.5.1:
```
Traceback (most recent call last):
  File "src/build/landmines.py", line 135, in <module>
    sys.exit(main())
  File "src/build/landmines.py", line 129, in main
    clobber_if_necessary(landmines)
  File "src/build/landmines.py", line 80, in clobber_if_necessary
    clobber.clobber(out_dir)
  File "/Users/aaclarke/dev/engine/src/build/clobber.py", line 99, in clobber
    delete_build_dir(path)
  File "/Users/aaclarke/dev/engine/src/build/clobber.py", line 60, in delete_build_dir
    shutil.rmtree(build_dir)
  File "/usr/local/Cellar/python@3.7/3.7.11/Frameworks/Python.framework/Versions/3.7/lib/python3.7/shutil.py", line 504, in rmtree
    onerror(os.path.islink, path, sys.exc_info())
  File "/usr/local/Cellar/python@3.7/3.7.11/Frameworks/Python.framework/Versions/3.7/lib/python3.7/shutil.py", line 502, in rmtree
    raise OSError("Cannot call rmtree on a symbolic link")
OSError: Cannot call rmtree on a symbolic link
```