Respect depfile specifications for compiled_action templates.
The more accurate way if implementing this template would be to use
`forward_variables_from` with an allowlist. But that functionality
was likely not available in GN when this template was written.

The case could be made that his should now be re-written to use
`forward_variables_from` to be as forward compatible as possible.

For now, I just needed to the template to respect the depfile
specification. This is to be used by the offline shader compiler which
is a C++ tool on the host (not vended).