Fix GLFW GN rule to add missing Mac dependencies.
Not including CoreVideo and IOKit here yields linker errors when the
final executable is linked. This required executable targets that have
no knowledge of CoreVideo or IOKit explicitly specify a dependency.

With this change, these dependencies will be implicitly added if the
target depends on GLFW.

The suppression -Wno-objc-multiple-method-names is necessary after
updating GLFW to the latest (which landing this patch will unblock).